+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-06-05 13:33:15
-// Commit: feat(state): integrate offline caching for state state
+// Updated at: 2026-06-05 16:10:29
+// Commit: fix(state): correct layout constraints and overflow in state
 
 class stateComponentState {
-  final String commitRef = '1764';
-  final String lastMessage = 'feat(state): integrate offline caching for state state';
-  final DateTime timestamp = DateTime.parse('2026-06-05T13:33:15+05:00');
+  final String commitRef = '1773';
+  final String lastMessage = 'fix(state): correct layout constraints and overflow in state';
+  final DateTime timestamp = DateTime.parse('2026-06-05T16:10:29+05:00');
 }